# ⚡ Optimize `KeyBytes` derivation