--- a/CHANGES.md
+++ b/CHANGES.md
@@ -10,6 +10,17 @@ To be released.
 
 ### Backward-incompatible API changes
 
+ -  Changed `StateStoreExtensions.Commit()` to accept
+    `IImmutableDictionary<KeyBytes, IValue>` instead of
+    `IImmutableDictionary<string, IValue>`.  [[#3321]]
+ -  Added `StateStoreExtensions.GetStates(IStateStore, HashDigest<SHA256>,
+    IReadOnlyList<KeyBytes>)` method.  [[#3321]]
+ -  Changed `AccountDeltaExtensions.ToRawDelta()` to return
+    `IImmutableDictionary<KeyBytes, IValue>` instead of
+    `IImmutableDictionary<string, IValue>`.  [[#3321]]
+ -  Changed `ActionEvaluatorExtensions.GetRawTotalDelta()` to return
+    `IImmutableDictionary<KeyBytes, IValue>` instead of
+    `IImmutableDictionary<string, IValue>`.  [[#3321]]
  -  Removed `EnumerableMeasurement` class.  [[#3325]]
  -  Removed `KeyValueExtensions` class.  [[#3325]]
 
@@ -21,12 +32,16 @@ To be released.
 
 ### Behavioral changes
 
+ -  Optimized read and write access to `IStateStore` both for memory and speed.
+    [[#3321]]
+
 ### Bug fixes
 
 ### Dependencies
 
 ### CLI tools
 
+[#3321]: https://github.com/planetarium/libplanet/pull/3321
 [#3325]: https://github.com/planetarium/libplanet/pull/3325