--- a/Libplanet.Action/State/AccountDeltaExtensions.cs
+++ b/Libplanet.Action/State/AccountDeltaExtensions.cs
@@ -5,6 +5,7 @@ using System.Numerics;
 using Bencodex.Types;
 using Libplanet.Crypto;
 using Libplanet.Store;
+using Libplanet.Store.Trie;
 using Libplanet.Types.Assets;
 using Libplanet.Types.Consensus;
 using static Libplanet.Action.State.KeyConverters;
@@ -49,19 +50,19 @@ namespace Libplanet.Action.State
         /// <param name="delta">The <see cref="IAccountDelta"/> to convert.</param>
         /// <returns>A raw dictionary representation of <see cref="IAccountDelta"/> to write
         /// to an <see cref="IStateStore"/>.</returns>
-        public static IImmutableDictionary<string, IValue> ToRawDelta(this IAccountDelta delta)
+        public static IImmutableDictionary<KeyBytes, IValue> ToRawDelta(this IAccountDelta delta)
         {
             var rawStates = delta.States.Select(
-                kv => new KeyValuePair<string, IValue>(
+                kv => new KeyValuePair<KeyBytes, IValue>(
                     ToStateKey(kv.Key), kv.Value));
             var rawFungibles = delta.Fungibles.Select(
-                kv => new KeyValuePair<string, IValue>(
+                kv => new KeyValuePair<KeyBytes, IValue>(
                     ToFungibleAssetKey(kv.Key), new Integer(kv.Value)));
             var rawTotalSupplies = delta.TotalSupplies.Select(
-                kv => new KeyValuePair<string, IValue>(
+                kv => new KeyValuePair<KeyBytes, IValue>(
                     ToTotalSupplyKey(kv.Key), new Integer(kv.Value)));
 
-            var rawDelta = ImmutableDictionary<string, IValue>.Empty;
+            var rawDelta = ImmutableDictionary<KeyBytes, IValue>.Empty;
             rawDelta = rawDelta.SetItems(rawStates.Concat(rawFungibles).Concat(rawTotalSupplies));
             return delta.ValidatorSet is { } validatorSet
                 ? rawDelta.SetItem(ValidatorSetKey, validatorSet.Bencoded)