--- a/Libplanet.Action/State/AccountStateDeltaExtensions.cs
+++ b/Libplanet.Action/State/AccountStateDeltaExtensions.cs
@@ -24,43 +24,5 @@ namespace Libplanet.Action.State
             stateDelta.Delta.TotalSupplies.ToImmutableDictionary(
                 kv => kv.Key,
                 kv => FungibleAssetValue.FromRawValue(kv.Key, kv.Value));
-
-        internal static IImmutableDictionary<string, IValue?> GetUpdatedRawStates(
-            this IAccountStateDelta stateDelta)
-        {
-            var dict = stateDelta.GetUpdatedStates()
-                .Select(
-                    pair =>
-                        new KeyValuePair<string, IValue?>(
-                            ToStateKey(pair.Key),
-                            pair.Value))
-                .Union(
-                    stateDelta.GetUpdatedBalances()
-                        .Select(
-                            pair =>
-                                new KeyValuePair<string, IValue?>(
-                                    ToFungibleAssetKey(pair.Key),
-                                    (Integer)pair.Value.RawValue)))
-                .Union(
-                    stateDelta.GetUpdatedTotalSupplies()
-                        .Select(
-                            pair =>
-                                new KeyValuePair<string, IValue?>(
-                                    ToTotalSupplyKey(pair.Key),
-                                    (Integer)pair.Value.RawValue)));
-
-            if (stateDelta.GetValidatorSet().Validators.Any())
-            {
-                dict = dict.Union(
-                    new[]
-                    {
-                        new KeyValuePair<string, IValue?>(
-                            ValidatorSetKey,
-                            stateDelta.GetValidatorSet().Bencoded),
-                    });
-            }
-
-            return dict.ToImmutableDictionary();
-        }
     }
 }