Tested with 1,000,000 batch.
By "bypass" I mean direct `byte[]` to `byte[]` conversion without obtaining `string` representation in the middle.
Speed-wise not that great, but still reduces quite a bit in memory usage.

|                 Method |       Mean |      Error |     StdDev | Allocated |
|----------------------- |-----------:|-----------:|-----------:|----------:|
|             ToStateKey | 321.206 ms |  6.4224 ms | 13.8248 ms | 312.81 MB |
|       ToStateKeyBypass | 262.234 ms |  5.0044 ms |  5.7631 ms |  129.7 MB |
|               ToFAVKey | 536.196 ms | 10.3974 ms | 15.2403 ms | 785.83 MB |
|         ToFAVKeyBypass | 430.898 ms |  8.5690 ms | 19.5161 ms | 221.25 MB |
|       ToTotalSupplyKey | 408.493 ms |  7.9977 ms | 21.6222 ms | 427.25 MB |
| ToTotalSupplyKeyBypass | 312.920 ms |  6.1483 ms | 11.2424 ms | 144.96 MB |